requirements.txt
flask
scikit-learn
python-telegram-bot
waitress
twilio
python-dotenv